Creating xml file for start of project
The downloaded files were osm.pbf which were converted to .osm using the converter provided in the executables. Then, the filter was used to create the file which only has the nodes/ways which are useful for our project.

The files that were used were downloaded from OpenStreetMap, more precisely Geofabrik downloads, we used the latest map for Macedonia in 2023.